# static/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Whiteflag API",
        "version": "1.0.0-alpha",
...
        "x-logo": {
            "url": "wflogo.png",
            "altText": "Whiteflag"
        }
    },
...
    "servers": [
        {
            "url": "http://{hostname}:{port}/",
            "variables": {
                "hostname": {
                    "default": "localhost",
                    "description": "Configurable hostname"
                },
                "port": {
                    "default": "5746",
                    "description": "Configurable port number"
                }
            }
        }
    ],
    "security": [
        {
            "description": "Basic HTTP authorization",
            "type": "http",
            "scheme": "basic"
        }
    ],
    "paths": {
...
        "/originators/{address}": {
            "get": {
                "tags": [ "Originators", "State" ],
                "description": "Returns the details of the specified Whiteflag originator.",
                "operationId": "getOriginator",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/originatorState"
                    },
                    "404": {
                        "$ref": "#/components/responses/error404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    }
                }
            }
        },
        "/originators/{address}/psk/{account}": {
            "get": {
                "tags": [ "Originators", "State" ],
                "description": "Checks for existence of a pre-shared key for the specified Whiteflag originator to be used with the specified blockchain account. This operation may be disabled in the configuration.",
                "operationId": "getPreSharedKey",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/originatorPsk"
                    },
                    "404": {
                        "$ref": "#/components/responses/error404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    }
                }
            },
            "put": {
                "tags": [ "Originators", "State" ],
                "description": "Stores or updates a pre-shared key for the specified Whiteflag originator to be used with the specified blockchain account. This operation may be disabled in the configuration.",
                "operationId": "storePreSharedKey",
                "requestBody": {
                    "$ref": "#/components/requestBodies/storePreSharedKey"
                },
                "responses": {
                    "202": {
                        "$ref": "#/components/responses/originatorPskPut"
                    },
                    "404": {
                        "$ref": "#/components/responses/error404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    }
                }
            },
            "delete": {
                "tags": [ "Originators", "State" ],
                "description": "Deletes a pre-shared key for the specified Whiteflag originator to be used with the specified blockchain account. This operation may be disabled in the configuration.",
                "operationId": "deletePreSharedKey",
                "responses": {
                    "202": {
                        "$ref": "#/components/responses/originatorPskDeleted"
                    },
                    "404": {
                        "$ref": "#/components/responses/error404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "schemas": {
...
            "responseBodyMetaObject": {
                "title": "API response metadata",
                "description": "Meta data about the processing of the request",
                "type": "object",
                "properties": {
                    "additionalProperties": true,
                    "operationId": {
                        "description": "Unique string defined by the OpenAPI defintion used to identify the operation",
                        "type": "string"
                    },
                    "request": {
                        "description": "Details of the request on the endpoint",
                        "type": "object",
                        "properties": {
                            "client": {
                                "description": "The ip address of the client that made the request",
                                "type": "string"
                            },
                            "method": {
                                "description": "The HTTP method used to make the request",
                                "type": "string"
                            },
                            "endpoint": {
                                "description": "The endpoint on which the request has been made",
                                "type": "string"
                            }
                        }
                    },
                    "info": {
                        "description": "Information about how the data was processed",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "warnings": {
                        "description": "Warnings generated when processing the request",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "errors": {
                        "description": "Errors generated when processing the request",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "responseBodyErrorObject": {
                "title": "API response errors",
                "description": "Errors describing why the request could not succesfully be fulfilled",
                "type": "array",
                "items": {
                    "type": "string"
                }
            },
            "responseBodyErrors": {
                "type": "object",
                "properties": {
                    "meta": {
                        "$ref": "#/components/schemas/responseBodyMetaObject"
                    },
                    "errors": {
                        "$ref": "#/components/schemas/responseBodyErrorObject"
                    }
                }
            }
        },
        "requestBodies": {
...
            "storePreSharedKey": {
                "description": "Request to store a pre-shared key for an originator",
                "content": {
                    "application/json": {
                        "schema": {
                            "description": "Pre-shared key data object",
                            "type": "object",
                            "properties": {
                                "preSharedKey": {
                                    "description": "A pre-shared key in raw hexadecimal format",
                                    "type": "string",
                                    "pattern": "[a-fA-F0-9]"
                                }
                            }
                        }
                    }
                }
            }
        },
        "responses": {
...
            "error404": {
                "description": "Requested resource not found or no data available",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseBodyErrors"
                        }
                    }
                }
            },
            "error500": {
                "description": "Internal error preventing the API to process the request",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseBodyErrors"
                        }
                    }
                }
            },
...
            "originatorState": {
                "description": "Successfully retrieved and returning known Whiteflag originators",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta": {
                                    "$ref": "#/components/schemas/responseBodyMetaObject"
                                },
                                "data": {
                                    "$ref": "protocol/state.schema.json#/definitions/originatorObject"
                                }
                            }
                        }
                    }
                }
            },
            "originatorPsk": {
                "description": "A pre-shared key for the Whiteflag originator is available",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta": {
                                    "$ref": "#/components/schemas/responseBodyMetaObject"
                                },
                                "data": {
                                    "$ref": "protocol/state.schema.json#/definitions/originatorObject"
                                }
                            }
                        }
                    }
                }
            },
            "originatorPskPut": {
                "description": "Successfully accepted new or updated pre-shared key for the Whiteflag originator",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta": {
                                    "$ref": "#/components/schemas/responseBodyMetaObject"
                                },
                                "data": {
                                    "$ref": "protocol/state.schema.json#/definitions/originatorObject"
                                }
                            }
                        }
                    }
                }
            },
            "originatorPskDeleted": {
                "description": "Successfully accepted the delete request for the pre-shared key for the Whiteflag originators",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta": {
                                    "$ref": "#/components/schemas/responseBodyMetaObject"
                                },
                                "data": {
                                    "$ref": "protocol/state.schema.json#/definitions/originatorObject"
                                }
                            }
                        }
                    }
                }
            },
...
        }
    }
}